@@ -1,13 +1,13 @@
 ext.versions = [
-        minSdk            : 13,
-        compileSdk        : 26,
+        minSdk            : 21,
+        compileSdk        : 32,
         buildTools        : '26.0.1',
         publishVersion    : '0.9.4.6',
         publishVersionCode: 176,
-        gradlePlugin      : '3.0.0-beta2',
+        gradlePlugin      : '7.4.2',
         spotlessPlugin    : '3.4.0',
 
         supportLib        : '25.3.1',
         mdProgressBar     : '1.4.1',
-        butterKnife       : '8.6.0'
+        butterKnife       : '10.2.3'
 ]
