@@ -5,8 +5,8 @@ import android.content.res.TypedArray;
 import android.content.res.XmlResourceParser;
 import android.preference.Preference;
 import android.preference.PreferenceManager;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.util.AttributeSet;
 import com.afollestad.materialdialogs.commons.R;
 import java.lang.reflect.Method;
