@@ -2,7 +2,7 @@ package com.afollestad.materialdialogs.util;
 
 import android.content.Context;
 import android.graphics.Typeface;
-import android.support.v4.util.SimpleArrayMap;
+import androidx.collection.SimpleArrayMap;
 
 /*
     Each call to Typeface.createFromAsset will load a new instance of the typeface into memory,
