@@ -2,8 +2,8 @@ package com.afollestad.materialdialogssample;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
-import android.support.annotation.ArrayRes;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.ArrayRes;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
