@@ -3,7 +3,7 @@ package com.afollestad.materialdialogssample;
 import android.annotation.SuppressLint;
 import android.os.Bundle;
 import android.preference.PreferenceFragment;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.MenuItem;
 
 @SuppressLint("NewApi")
